backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
fastapi==0.136.0
h11==0.16.0
idna==3.11
invoke==3.0.3
lxml==6.0.4
markdown-it-py==4.0.0
mdurl==0.1.2
paramiko==4.0.0
proxmoxer==2.3.0
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyNaCl==1.6.2
python-dotenv==1.2.2
requests==2.33.1
rich==15.0.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
